requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
openpyxl